+++ aiomisc/timeout.py
@@ -1,4 +1,5 @@
 import asyncio
+import inspect
 from collections.abc import Callable, Coroutine
 from functools import wraps
 from typing import Any, ParamSpec, TypeVar
@@ -16,7 +17,7 @@ def timeout(
     def decorator(
         func: Callable[P, Coroutine[Any, Any, T]],
     ) -> Callable[P, Coroutine[Any, Any, T]]:
-        if not asyncio.iscoroutinefunction(func):
+        if not inspect.iscoroutinefunction(func):
             raise TypeError("Function is not a coroutine function")
 
         @wraps(func)